Use Scanner.nextLine() instead of Scanner.next().
The latter has the following problem. The code...

println("Please input a line of text: ");
String str = readLine();
int i = str.length();
println(i);

...results in...

Please input a line of text:
This is a test
4

...because next() only takes the string up the blank, not up to the
CR/LF.

Not properly tested yet. Hope this does not blow up because of some
unexpected side-effect of nextLine().