Kind of fixes Issue #7.
I was so fed up with creting ZIPs manually that I just wrote a script
because I could not face doing it once more.

It did not take more than 30 minutes, including generation of
sub-scripts, sample programs, and error-checking. Should have done it
long ago.

Issue #7 is about using Maven to do this, but this scratches my itch
completely so I may as well close the issue. It is true that a shell
script only works in unix and not in windows, but that is what cygwin
is for (and how many JavaDecaf developers will deploy from Windows,
anyway?).